+++ src/main/scala/advent/day3.scala
@@ -5,7 +5,7 @@ trait Day3Common:
 
   val parser: Parser[Vector[String]] = PBasic(RString())
 
-  def extractMulResults(input: String): Int =
+  def extractMulResults(input: String): Long =
     val mulPattern = "mul\\((\\d+),(\\d+)\\)".r
 
     val matches = mulPattern.findAllMatchIn(input)
@@ -13,13 +13,13 @@ trait Day3Common:
     matches.map(m => m.group(1).toInt * m.group(2).toInt).sum
 
 given day3part1Solution: Solver[3, 1] = new Solver[3, 1] with Day3Common:
-  override def solve(input: Vector[String]): Int =
+  override def solve(input: Vector[String]): Long =
     input.map(extractMulResults).sum
 
 given day3part2Solution: Solver[3, 2] = new Solver[3, 2] with Day3Common:
-  override def solve(input: Vector[String]): Int =
+  override def solve(input: Vector[String]): Long =
     input
-      .foldLeft((0, true)) { case ((sum, isEnabled), line) =>
+      .foldLeft((0L, true)) { case ((sum, isEnabled), line) =>
         line.split("don't\\(\\)") match
           case Array() => throw InvalidLineError(line)
           case Array(single) =>
